docuagent-backend/requirements.txt
fastapi
uvicorn
autogen
azure-storage-blob
azure-cosmos
PyMuPDF
python-multipart
python-dotenv
pyautogen
python-docx